tui: split deck_ui.rs's test module out of the god-file
`stella-tui/src/deck_ui.rs` was 6,884 lines — over 4x the 1,500-line ratchet
`scripts/check-file-size.sh` enforces, and the single worst file in the tree.
About 2,960 of those lines were its own `#[cfg(test)] mod tests`, which is the
half that can move with no behavior risk at all: no production code changes,
no signature changes, no visibility changes.

deck_ui.rs is now 3,926 lines. The tests live in `deck_ui/tests.rs` (shared
fixtures) plus five submodules split on the section comments the test module
already carried:

  deck_ui/tests.rs          1,110   shared fixtures + key/ingest tests
  deck_ui/tests/queue.rs      458   queue editor, slash popup, MCP tab
  deck_ui/tests/issues.rs     415   ISSUES tab
  deck_ui/tests/graph.rs      407   graph picker, installed-agents browser
  deck_ui/tests/skills.rs     397   SKILLS tab
  deck_ui/tests/help.rs       196   help + inspect/context overlays

Every new file is under the 1,500-line limit, which the ratchet requires of
files that are not grandfathered. The layout matches what the crate already
uses for `fleet_dashboard.rs`/`fleet_dashboard/tests.rs` and
`deck_render.rs`/`deck_render/tests.rs`, so this is the established
convention rather than a new one.

Section-local fixtures travelled with their section; only `model_with_queue`
turned out to be shared across the boundary and was hoisted into
`tests.rs`. The compiler caught it — nothing here relies on my reading of
which helper is used where.

123 tests pass, the same 123 that pass on `main` — verified by running the
`deck_ui` filter on both sides of the change. Test count identical, names
unchanged apart from the submodule path segment.

The baseline entry is tightened by hand to 3,926 rather than by running
`make file-size-update`, which regenerates every ceiling at current size.
A wholesale regeneration would tighten unrelated files' ceilings to their
exact current length and break the other PRs in flight against them; only
the line this change actually affects should move.

Not done here, and still open on #458: the remaining 3,926 lines of
`deck_ui.rs` code, and `command_deck.rs` (4,489). command_deck.rs already
has its tests extracted, so its bulk is `run_deck_session`, whose borrow
structure across six live mpsc channels is exactly the risky half the audit
declined — that one needs its own PR and its own verification, not a
drive-by at the end of this one.

Refs #458, #620
