+++ stella-tui/src/deck_ui.rs
@@ -3923,2962 +3923,4 @@ fn cycle_filter(model: &WorkspaceModel, current: Option<&str>) -> Option<AgentId
 }
 
 #[cfg(test)]
-#[allow(clippy::field_reassign_with_default)]
-mod tests {
-    use super::*;
-    use crate::envelope::AgentMeta;
-    use stella_protocol::AgentEvent;
-
-    fn key(code: KeyCode) -> KeyEvent {
-        KeyEvent::new(code, KeyModifiers::NONE)
-    }
-    fn ch(c: char) -> KeyEvent {
-        key(KeyCode::Char(c))
-    }
-    fn ctrl(c: char) -> KeyEvent {
-        KeyEvent::new(KeyCode::Char(c), KeyModifiers::CONTROL)
-    }
-    /// The newline chord — `⌘⏎` as the kitty keyboard protocol reports it
-    /// (a modified Enter inserts a line break; a bare Enter submits).
-    fn cmd_enter() -> KeyEvent {
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::SUPER)
-    }
-    fn alt(c: char) -> KeyEvent {
-        KeyEvent::new(KeyCode::Char(c), KeyModifiers::ALT)
-    }
-
-    fn model_with(ids: &[&str]) -> WorkspaceModel {
-        let mut m = WorkspaceModel::new();
-        for id in ids {
-            m.apply_inbound(&Inbound::Register(AgentMeta::new(*id, *id, 0)));
-        }
-        m
-    }
-
-    /// Push one tool call + multi-line result onto `agent`'s transcript.
-    fn with_tool_exchange(m: &mut WorkspaceModel, agent: &str) {
-        use stella_protocol::{AgentEvent, ToolCall, ToolOutput};
-        m.apply_inbound(&Inbound::Event {
-            agent: agent.into(),
-            event: AgentEvent::ToolStart {
-                call: ToolCall {
-                    call_id: "c1".into(),
-                    name: "read_file".into(),
-                    input: serde_json::json!({ "path": "src/main.rs" }),
-                },
-            },
-        });
-        m.apply_inbound(&Inbound::Event {
-            agent: agent.into(),
-            event: AgentEvent::ToolResult {
-                call_id: "c1".into(),
-                output: ToolOutput::Ok {
-                    content: "line one\nline two\nline three".into(),
-                },
-                duration_ms: 7,
-                speculated: false,
-            },
-        });
-    }
-
-    #[test]
-    fn up_selects_the_last_message_and_ctrl_o_toggles_it() {
-        let mut model = model_with(&["lead"]);
-        with_tool_exchange(&mut model, "lead");
-        let mut ui = ready_ui();
-
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(
-            ui.session_selected,
-            Some(1),
-            "up highlights the newest message"
-        );
-
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(
-            ui.expanded.get("lead").is_some_and(|set| set.contains(&1)),
-            "ctrl+o expands the highlighted message"
-        );
-        let rev = ui.expanded_rev;
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(
-            ui.expanded.get("lead").is_none_or(|set| !set.contains(&1)),
-            "a second ctrl+o collapses it again"
-        );
-        assert!(
-            ui.expanded_rev > rev,
-            "each toggle invalidates the fold cache"
-        );
-    }
-
-    #[test]
-    fn ctrl_o_with_no_selection_toggles_the_expand_all_overlay() {
-        let mut model = model_with(&["lead"]);
-        with_tool_exchange(&mut model, "lead");
-        let mut ui = ready_ui();
-
-        // First press (no selection): every expandable message opens at once —
-        // no per-entry set is touched.
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(ui.transcript_expand_all, "expand-all overlay on");
-        assert!(
-            ui.expanded.get("lead").is_none_or(|set| set.is_empty()),
-            "the overlay does not write the per-entry sets"
-        );
-        // Second press: everything closes again — ctrl+o is its own way out.
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(!ui.transcript_expand_all, "ctrl+o again collapses");
-    }
-
-    #[test]
-    fn esc_collapses_the_expand_all_overlay_before_stopping_the_turn() {
-        let mut model = model_with(&["lead"]);
-        with_tool_exchange(&mut model, "lead"); // events flip the agent to Running
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(ui.transcript_expand_all);
-
-        // Esc's first job here is the overlay — NOT cancelling the running
-        // turn (precedence rule 8 beats rules 9–10).
-        let action = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert!(!ui.transcript_expand_all, "esc is a graceful way out");
-        assert!(
-            matches!(action, DeckAction::Handled),
-            "the overlay-collapsing esc must not reach the turn-stop rules"
-        );
-        // With the overlay gone, the next Esc resumes normal duty (stop the
-        // in-flight turn).
-        let action = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert!(
-            matches!(action, DeckAction::Send(WorkspaceInput::Control { .. })),
-            "esc after the overlay closes stops the turn as before"
-        );
-    }
-
-    #[test]
-    fn ctrl_o_on_a_highlight_peels_one_row_out_of_the_expand_all_overlay() {
-        let mut model = model_with(&["lead"]);
-        with_tool_exchange(&mut model, "lead"); // entries 0 (call) + 1 (result)
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('o'), &model, &mut ui); // overlay on
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui); // highlight entry 1
-
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(
-            !ui.transcript_expand_all,
-            "the overlay materializes into per-entry expansions"
-        );
-        let set = ui.expanded.get("lead").expect("materialized set");
-        assert!(
-            set.contains(&0) && !set.contains(&1),
-            "the highlighted row collapsed; the rest stay open: {set:?}"
-        );
-    }
-
-    #[test]
-    fn bracket_jumps_reach_both_ends_of_the_transcript() {
-        let mut model = model_with(&["lead"]);
-        with_tool_exchange(&mut model, "lead");
-        let mut ui = ready_ui();
-        // A scrollable transcript (metrics as the render pass would set them).
-        ui.metrics.session_total = 100;
-        ui.metrics.session_height = 10;
-
-        // ⌘/⌃ [ pins the window to the very beginning of the session…
-        handle_deck_key(ctrl('['), &model, &mut ui);
-        assert!(!ui.session_scroll.follow);
-        assert_eq!(ui.session_scroll.window(100, 10), 0..10, "jumped to start");
-
-        // …and ⌘/⌃ ] returns to the end, re-arming tail-follow. Both also
-        // drop any highlight so the pinned selection can't yank the view back.
-        let cmd_close = KeyEvent::new(KeyCode::Char(']'), KeyModifiers::SUPER);
-        handle_deck_key(cmd_close, &model, &mut ui);
-        assert!(ui.session_scroll.follow, "jumped to end = follow the tail");
-        assert_eq!(ui.session_selected, None);
-    }
-
-    #[test]
-    fn down_past_the_last_message_clears_selection_and_rearms_follow() {
-        let mut model = model_with(&["lead"]);
-        with_tool_exchange(&mut model, "lead");
-        let mut ui = ready_ui();
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.session_selected, None, "down past the tail deselects");
-        assert!(ui.session_scroll.follow, "…and re-arms tail-follow");
-    }
-
-    #[test]
-    fn ctrl_o_on_a_non_expandable_selection_is_a_no_op() {
-        let mut model = model_with(&["lead"]);
-        // A plain text message — `entry_lines` ignores the expanded flag for
-        // it, so ctrl+o has nothing to toggle.
-        model.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::Text {
-                delta: "hello".into(),
-            },
-        });
-        let mut ui = ready_ui();
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(ui.session_selected, Some(0));
-
-        let rev = ui.expanded_rev;
-        handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert!(
-            ui.expanded.get("lead").is_none_or(|set| set.is_empty()),
-            "nothing marked expanded"
-        );
-        assert_eq!(
-            ui.expanded_rev, rev,
-            "a no-op press must not invalidate the settled fold cache"
-        );
-    }
-
-    #[test]
-    fn switching_focus_drops_the_session_selection() {
-        let mut model = model_with(&["lead", "sub"]);
-        with_tool_exchange(&mut model, "lead");
-        let mut ui = ready_ui();
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert!(ui.session_selected.is_some());
-
-        // Focus another agent from the Agents tab: the selection indexes the
-        // *previous* agent's transcript and must not carry across.
-        ui.tab = DeckTab::Agents;
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.focused, 1);
-        assert_eq!(
-            ui.session_selected, None,
-            "selection cleared on focus change"
-        );
-        assert!(ui.session_scroll.follow, "…and tail-follow re-arms");
-    }
-
-    #[test]
-    fn deregister_keeps_focus_in_bounds_and_on_the_same_agent() {
-        let mut model = model_with(&["lead", "req:1", "req:2"]);
-        let mut ui = ready_ui();
-        ui.focused = 2; // req:2
-
-        // An EARLIER row vanishing shifts indexes down: the focused AGENT
-        // stays focused at its new index.
-        ingest_inbound(
-            &Inbound::Deregister {
-                agent: "req:1".into(),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert_eq!(ui.focused, 1);
-        assert_eq!(model.agents[ui.focused].meta.id, "req:2");
-
-        // The focused LAST row vanishing clamps focus back into range.
-        ingest_inbound(
-            &Inbound::Deregister {
-                agent: "req:2".into(),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert_eq!(ui.focused, 0, "focus stays in bounds");
-        assert_eq!(model.agents[ui.focused].meta.id, "lead");
-    }
-
-    #[test]
-    fn deregister_of_the_focused_row_drops_the_stale_selection() {
-        let mut model = model_with(&["lead", "req:1", "req:2"]);
-        with_tool_exchange(&mut model, "req:1");
-        with_tool_exchange(&mut model, "req:2");
-        let mut ui = ready_ui();
-        ui.focused = 1; // req:1
-        ui.session_selected = Some(1); // a row of req:1's transcript
-
-        ingest_inbound(
-            &Inbound::Deregister {
-                agent: "req:1".into(),
-            },
-            &mut model,
-            &mut ui,
-        );
-        // The successor (req:2) slides into the focused index…
-        assert_eq!(ui.focused, 1);
-        assert_eq!(model.agents[1].meta.id, "req:2");
-        // …but the selection indexed the REMOVED agent's transcript, so it
-        // must not re-attach to the successor's (which is long enough that
-        // range-clamping alone would have kept it).
-        assert_eq!(ui.session_selected, None);
-        assert!(ui.session_scroll.follow, "tail-follow re-arms");
-    }
-
-    fn ready_ui() -> DeckUi {
-        let mut ui = DeckUi::default();
-        ui.splash.skip(); // past the splash for interaction tests
-        ui
-    }
-
-    fn session_info(id: &str) -> crate::envelope::SessionInfo {
-        crate::envelope::SessionInfo {
-            id: id.into(),
-            title: format!("title for {id}"),
-            summary: String::new(),
-            workspace: "/tmp/w".into(),
-            phase: crate::envelope::SessionPhase::Complete,
-            started_ms: 0,
-            updated_ms: 0,
-            mine: false,
-            resumable: false,
-        }
-    }
-
-    fn notification(
-        id: &str,
-        read: bool,
-        session: Option<&str>,
-    ) -> crate::envelope::NotificationInfo {
-        crate::envelope::NotificationInfo {
-            id: id.into(),
-            title: "a title".into(),
-            body: "a body".into(),
-            source: String::new(),
-            created_ms: 0,
-            read,
-            session_id: session.map(str::to_string),
-        }
-    }
-
-    #[test]
-    fn sessions_overlay_enter_opens_the_selected_session_and_closes() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.sessions_open = true;
-        ui.sessions = vec![session_info("ses-1"), session_info("ses-2")];
-        ui.sessions_sel = 1;
-
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::SessionOpen { id: "ses-2".into() }),
-            "⏎ opens (replays) the selected session"
-        );
-        assert!(!ui.sessions_open, "the overlay closes on open");
-    }
-
-    #[test]
-    fn sessions_overlay_enter_with_no_rows_is_a_no_op() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.sessions_open = true; // registry snapshot empty
-
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled, "nothing to open");
-        assert!(ui.sessions_open, "the overlay stays up (Esc closes it)");
-    }
-
-    #[test]
-    fn inbox_enter_on_a_linked_notification_marks_read_and_opens_the_session() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.inbox_open = true;
-        ui.notifications = vec![notification("n1", false, Some("ses-9"))];
-
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::NotificationRead { id: "n1".into() }),
-            "the read goes out as the key's action"
-        );
-        assert_eq!(
-            ui.pending_inputs,
-            vec![WorkspaceInput::SessionOpen { id: "ses-9".into() }],
-            "…and the open rides pending_inputs right behind it"
-        );
-        assert!(!ui.inbox_open, "the overlay closes when a session opens");
-    }
-
-    #[test]
-    fn inbox_enter_on_an_already_read_linked_notification_just_opens() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.inbox_open = true;
-        ui.notifications = vec![notification("n1", true, Some("ses-9"))];
-
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::SessionOpen { id: "ses-9".into() }),
-            "already read — no second NotificationRead, just the open"
-        );
-        assert!(ui.pending_inputs.is_empty());
-        assert!(!ui.inbox_open);
-    }
-
-    #[test]
-    fn inbox_enter_without_a_session_link_keeps_the_mark_read_behavior() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.inbox_open = true;
-        ui.notifications = vec![notification("n1", false, None)];
-
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::NotificationRead { id: "n1".into() }),
-            "unlinked ⏎ is exactly the old mark-read"
-        );
-        assert!(ui.pending_inputs.is_empty(), "no session to open");
-        assert!(ui.inbox_open, "the overlay stays open, as before");
-
-        // Once read, ⏎ on an unlinked notification is a no-op.
-        ui.notifications = vec![notification("n1", true, None)];
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled);
-        assert!(ui.inbox_open);
-    }
-
-    #[test]
-    fn inbox_space_only_marks_read_and_never_opens() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.inbox_open = true;
-        ui.notifications = vec![notification("n1", false, Some("ses-9"))];
-
-        let action = handle_deck_key(key(KeyCode::Char(' ')), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::NotificationRead { id: "n1".into() }),
-            "␣ keeps its plain mark-read meaning"
-        );
-        assert!(ui.pending_inputs.is_empty(), "␣ never opens the session");
-        assert!(ui.inbox_open, "␣ never closes the overlay");
-    }
-
-    #[test]
-    fn eviction_clamps_the_selection_and_drops_stale_expansions() {
-        use crate::model::MAX_TRANSCRIPT_ENTRIES;
-        let mut model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        let retry = |i: usize| Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::Retry {
-                attempt: i as u32,
-                reason: "r".into(),
-            },
-        };
-        // Grow to just under the cap, then highlight + expand near the tail.
-        for i in 0..(MAX_TRANSCRIPT_ENTRIES - 1) {
-            ingest_inbound(&retry(i), &mut model, &mut ui);
-        }
-        ui.session_selected = Some(MAX_TRANSCRIPT_ENTRIES - 2);
-        toggle_expanded(&mut ui, "lead", MAX_TRANSCRIPT_ENTRIES - 2);
-        let rev = ui.expanded_rev;
-
-        // One more event crosses the cap: a chunk of the front evicts.
-        ingest_inbound(&retry(MAX_TRANSCRIPT_ENTRIES), &mut model, &mut ui);
-        let len = model.agents[0].model.transcript.len();
-        assert!(len < MAX_TRANSCRIPT_ENTRIES, "a chunk was evicted");
-        assert!(
-            ui.session_selected.is_some_and(|sel| sel < len),
-            "selection clamped into the retained window"
-        );
-        assert!(
-            !ui.expanded.contains_key("lead"),
-            "index-keyed expansions are stale once the front moved"
-        );
-        assert!(ui.expanded_rev > rev, "fold cache invalidated");
-    }
-
-    #[test]
-    fn any_key_dismisses_the_splash_first() {
-        let model = model_with(&["lead"]);
-        let mut ui = DeckUi::default(); // splash NOT skipped
-        assert!(!ui.splash.is_done());
-        assert_eq!(
-            handle_deck_key(ch('a'), &model, &mut ui),
-            DeckAction::Handled
-        );
-        assert!(ui.splash.is_done(), "first key skips the splash");
-        assert!(ui.composer.buffer().is_empty(), "and does not type");
-    }
-
-    #[test]
-    fn splash_cues_replay_and_release_the_cinematic() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = DeckUi::default();
-        ui.splash.skip(); // the deck is up; `/init` arrives later
-        assert!(ui.splash.is_done());
-
-        // Replay: a fresh held splash owns the frame again…
-        ingest_inbound(&Inbound::Splash(SplashCue::Replay), &mut model, &mut ui);
-        assert!(!ui.splash.is_done(), "replay restarts the cinematic");
-
-        // …and Release is what lets its timeline run out (exact timing is
-        // splash::tests territory — here we only pin that the cue routes).
-        ingest_inbound(&Inbound::Splash(SplashCue::Release), &mut model, &mut ui);
-        assert!(!ui.splash.is_done(), "the battle floor still plays out");
-    }
-
-    #[test]
-    fn no_anim_sessions_ignore_splash_replays() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = DeckUi::default();
-        ui.no_anim = true;
-        ui.splash.skip();
-        ingest_inbound(&Inbound::Splash(SplashCue::Replay), &mut model, &mut ui);
-        assert!(
-            ui.splash.is_done(),
-            "a no-anim session never replays the cinematic"
-        );
-    }
-
-    fn session_row(
-        id: &str,
-        phase: crate::envelope::SessionPhase,
-        mine: bool,
-        resumable: bool,
-    ) -> crate::envelope::SessionInfo {
-        crate::envelope::SessionInfo {
-            id: id.into(),
-            title: format!("ws: {id}"),
-            summary: String::new(),
-            workspace: "/w".into(),
-            phase,
-            started_ms: 0,
-            updated_ms: 0,
-            mine,
-            resumable,
-        }
-    }
-
-    #[test]
-    fn sessions_overlay_enter_resumes_resumable_rows_and_opens_the_rest() {
-        use crate::envelope::SessionPhase;
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.sessions_open = true;
-        ui.sessions = vec![
-            session_row("ses-mine", SessionPhase::InProgress, true, false),
-            session_row("ses-paused", SessionPhase::Paused, false, true),
-            session_row("ses-foreign", SessionPhase::Complete, false, false),
-        ];
-
-        // Grouped order: InProgress (mine) · Paused (resumable) · Complete.
-        // ⏎ on the resumable row navigates into it LIVE: the overlay closes
-        // and the driver is told to resume exactly that session.
-        ui.sessions_sel = 1;
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::SessionResume {
-                id: "ses-paused".into()
-            })
-        );
-        assert!(!ui.sessions_open, "the overlay closes on navigation");
-
-        // ⏎ on any non-resumable row — this deck's own included — opens a
-        // read-only replay instead (the `replay:<id>` lane).
-        for (sel, id) in [(0, "ses-mine"), (2, "ses-foreign")] {
-            ui.sessions_open = true;
-            ui.sessions_sel = sel;
-            assert_eq!(
-                handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-                DeckAction::Send(WorkspaceInput::SessionOpen { id: id.into() })
-            );
-            assert!(!ui.sessions_open, "the overlay closes on open too");
-        }
-    }
-
-    #[test]
-    fn paused_sessions_group_between_needs_input_and_cancelled() {
-        use crate::envelope::SessionPhase;
-        let mut ui = DeckUi::default();
-        ui.sessions = vec![
-            session_row("c", SessionPhase::Cancelled, false, true),
-            session_row("p", SessionPhase::Paused, false, true),
-            session_row("n", SessionPhase::NeedsInput, false, false),
-        ];
-        let order: Vec<&str> = grouped_session_rows(&ui)
-            .iter()
-            .map(|s| s.id.as_str())
-            .collect();
-        assert_eq!(order, ["n", "p", "c"]);
-    }
-
-    #[test]
-    fn only_tab_switches_tabs_and_digits_always_type() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        assert_eq!(ui.tab, DeckTab::Session);
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Agents);
-        handle_deck_key(key(KeyCode::BackTab), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Session);
-        // A digit with an empty composer starts the prompt — it never jumps
-        // to a tab, so prompts can begin with 1–5.
-        handle_deck_key(ch('3'), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Session, "digit typed, tab unchanged");
-        handle_deck_key(ch('h'), &model, &mut ui);
-        handle_deck_key(ch('2'), &model, &mut ui);
-        assert_eq!(ui.composer.buffer(), "3h2");
-    }
-
-    #[test]
-    fn bare_enter_always_enqueues_a_prompt_without_blocking() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "do the thing".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::Enqueue {
-                text: "do the thing".into()
-            })
-        );
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "composer clears after submit"
-        );
-    }
-
-    #[test]
-    fn a_modified_enter_inserts_a_line_break_preserved_through_submit() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "line one".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(
-            handle_deck_key(cmd_enter(), &model, &mut ui),
-            DeckAction::Handled,
-            "⌘⏎ is a line break, not a submit"
-        );
-        for c in "line two".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::Enqueue {
-                text: "line one\nline two".into()
-            }),
-            "the typed line break survives into the submitted prompt"
-        );
-    }
-
-    #[test]
-    fn plain_enter_on_a_blank_composer_inserts_nothing() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Graph; // a tab with no Enter binding of its own
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-            DeckAction::Ignored
-        );
-        assert!(ui.composer.buffer().is_empty(), "no stray leading newline");
-    }
-
-    #[test]
-    fn alt_brackets_jump_the_cursor_to_start_and_end() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "abc".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(ui.composer.cursor(), 3);
-        assert_eq!(
-            handle_deck_key(alt('['), &model, &mut ui),
-            DeckAction::Handled
-        );
-        assert_eq!(ui.composer.cursor(), 0, "⌥[ → before the first character");
-        assert_eq!(
-            handle_deck_key(alt(']'), &model, &mut ui),
-            DeckAction::Handled
-        );
-        assert_eq!(ui.composer.cursor(), 3, "⌥] → one past the last character");
-    }
-
-    #[test]
-    fn bare_enter_queues_and_a_modified_enter_inserts_a_break() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "hi".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let alt_enter = KeyEvent::new(KeyCode::Enter, KeyModifiers::ALT);
-        assert_eq!(
-            handle_deck_key(alt_enter, &model, &mut ui),
-            DeckAction::Handled,
-            "⌥⏎ inserts a line break"
-        );
-        assert_eq!(ui.composer.buffer(), "hi\n");
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::Enqueue {
-                text: "hi\n".into()
-            }),
-            "bare ⏎ queues (never blocks)"
-        );
-    }
-
-    #[test]
-    fn arrow_keys_edit_a_multiline_prompt_instead_of_scrolling() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "ab".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(cmd_enter(), &model, &mut ui); // ⌘⏎ inserts a line break
-        for c in "cd".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        // ↑ moves the cursor into the first line (not the session scroll,
-        // and NOT the queue editor — the composer is not empty).
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(ui.composer.cursor(), 2, "column kept on the line above");
-        handle_deck_key(key(KeyCode::Left), &model, &mut ui);
-        handle_deck_key(ch('X'), &model, &mut ui);
-        assert_eq!(ui.composer.buffer(), "aXb\ncd", "typed at the cursor");
-    }
-
-    #[test]
-    fn ctrl_c_quits_from_any_tab() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Graph;
-        assert_eq!(
-            handle_deck_key(ctrl('c'), &model, &mut ui),
-            DeckAction::Quit
-        );
-    }
-
-    /// A one-agent model whose lead is mid-turn (`Running`).
-    fn running_model() -> WorkspaceModel {
-        let mut m = model_with(&["lead"]);
-        m.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::Text {
-                delta: "working".into(),
-            },
-        });
-        m
-    }
-
-    /// The single-Esc outcome: a clean stop for the lead's in-flight turn.
-    fn stop_lead() -> DeckAction {
-        DeckAction::Send(WorkspaceInput::Control {
-            agent: "lead".into(),
-            control: AgentControl::Stop,
-        })
-    }
-
-    #[test]
-    fn esc_stops_a_running_turn_and_arms_the_double_press() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead()
-        );
-        assert!(
-            ui.esc_armed_at.is_some(),
-            "the stop arms the double-Esc window"
-        );
-    }
-
-    #[test]
-    fn esc_with_no_turn_running_stays_inert() {
-        let model = model_with(&["lead"]); // Queued — nothing in flight
-        let mut ui = ready_ui();
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Ignored,
-            "an idle Esc must not send a stray stop"
-        );
-        assert!(ui.esc_armed_at.is_none());
-    }
-
-    #[test]
-    fn a_typed_draft_survives_both_esc_forms() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        for c in "keep me".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        // The cursor lives in the global composer, so the stop fires even
-        // mid-draft — and must leave the draft untouched.
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead()
-        );
-        assert_eq!(ui.composer.buffer(), "keep me");
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::StopAndHold {
-                agent: "lead".into()
-            })
-        );
-        assert_eq!(
-            ui.composer.buffer(),
-            "keep me",
-            "neither cancel form clears what the user typed"
-        );
-    }
-
-    #[test]
-    fn double_esc_inside_the_window_escalates_to_stop_and_hold() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead()
-        );
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::StopAndHold {
-                agent: "lead".into()
-            })
-        );
-        assert!(
-            ui.dispatch_held,
-            "the deck now front-inserts its next submission"
-        );
-        assert!(
-            ui.esc_armed_at.is_none(),
-            "the pair resets after escalating"
-        );
-    }
-
-    #[test]
-    fn the_second_esc_fires_even_if_the_cancel_already_folded() {
-        // Between the two presses the first cancel's error event may fold
-        // (status `Failed`) before the auto-dispatched next prompt produces
-        // any event — the escalation must not be lost to that gap.
-        let mut model = running_model();
-        let mut ui = ready_ui();
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead()
-        );
-        model.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::Error {
-                message: "turn stopped by user".into(),
-                retryable: false,
-            },
-        });
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::StopAndHold {
-                agent: "lead".into()
-            })
-        );
-    }
-
-    #[test]
-    fn an_intervening_key_breaks_the_double_esc_pair() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        handle_deck_key(ch('x'), &model, &mut ui); // types into the composer
-        assert!(ui.esc_armed_at.is_none(), "any other key disarms");
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead(),
-            "the next Esc is a fresh single stop, not an escalation"
-        );
-    }
-
-    #[test]
-    fn a_stale_arm_outside_the_window_does_not_escalate() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        // Backdate the arm past the window. (If the monotonic clock is too
-        // young to backdate, `checked_sub` leaves it unarmed — which expects
-        // the same single-stop outcome.)
-        ui.esc_armed_at = std::time::Instant::now()
-            .checked_sub(ESC_DOUBLE_WINDOW + std::time::Duration::from_secs(1));
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead(),
-            "past the window, Esc is a single stop again"
-        );
-    }
-
-    #[test]
-    fn esc_dismisses_the_slash_popup_instead_of_stopping_the_turn() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![SlashCommand::new("/help", "help")];
-        handle_deck_key(ch('/'), &model, &mut ui);
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Handled,
-            "rule 4: the popup claims Esc — no stop is sent"
-        );
-        assert!(ui.esc_armed_at.is_none(), "a claimed Esc never arms");
-    }
-
-    #[test]
-    fn an_esc_claimed_by_the_queue_editor_breaks_the_pair_too() {
-        let mut model = model_with_queue(&["one"]);
-        model.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::Text { delta: "hi".into() },
-        });
-        let mut ui = ready_ui();
-        ui.queue_open = true;
-        ui.esc_armed_at = Some(std::time::Instant::now());
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Handled,
-            "rule 3: the queue editor claims Esc — no stop is sent"
-        );
-        assert!(!ui.queue_open, "…it closed the editor");
-        assert!(ui.esc_armed_at.is_none(), "…and broke the double-Esc pair");
-    }
-
-    #[test]
-    fn esc_still_aborts_a_pending_scope_review() {
-        let mut model = model_with(&["lead"]);
-        model.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::ScopeReview {
-                proposal: stella_protocol::ScopeProposal {
-                    summary: "big".into(),
-                    steps: vec![],
-                    estimated_files: 3,
-                    estimated_cost_usd: None,
-                },
-            },
-        });
-        let mut ui = ready_ui();
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::ToAgent {
-                agent: "lead".into(),
-                input: UserInput::ScopeDecision(ScopeDecision::Abort),
-            }),
-            "rule 5: the gate claims Esc — never a turn stop"
-        );
-    }
-
-    #[test]
-    fn esc_closes_an_open_diff_before_it_stops_the_turn() {
-        let model = running_model();
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Files;
-        ui.files_diff_open = true;
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Handled,
-            "rule 6: the diff overlay claims Esc"
-        );
-        assert!(!ui.files_diff_open);
-    }
-
-    #[test]
-    fn esc_clears_a_session_highlight_before_it_stops_the_turn() {
-        let mut model = running_model();
-        with_tool_exchange(&mut model, "lead");
-        let mut ui = ready_ui();
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert!(ui.session_selected.is_some());
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            DeckAction::Handled,
-            "rule 7: the highlight claims Esc first"
-        );
-        assert_eq!(ui.session_selected, None);
-        // With nothing left to claim it, the NEXT Esc stops the turn (and
-        // since the claimed Esc broke the pair, it is a single stop).
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-            stop_lead()
-        );
-    }
-
-    #[test]
-    fn the_first_submission_after_a_hold_enqueues_at_the_front() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.dispatch_held = true;
-        for c in "urgent fix".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::EnqueueFront {
-                text: "urgent fix".into()
-            }),
-            "the held submission jumps the queue"
-        );
-        assert!(!ui.dispatch_held, "the submission releases the hold");
-        for c in "later".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::Enqueue {
-                text: "later".into()
-            }),
-            "after the hold clears, submissions append as usual"
-        );
-    }
-
-    #[test]
-    fn agents_tab_controls_fire_only_when_composer_empty() {
-        let model = model_with(&["lead", "sub"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Agents;
-        ui.focused = 1;
-        // 's' with empty composer → Stop control for the focused agent.
-        assert_eq!(
-            handle_deck_key(ch('s'), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::Control {
-                agent: "sub".into(),
-                control: AgentControl::Stop,
-            })
-        );
-        // With text typed, 's' types instead.
-        handle_deck_key(ch('h'), &model, &mut ui);
-        handle_deck_key(ch('s'), &model, &mut ui);
-        assert_eq!(ui.composer.buffer(), "hs");
-    }
-
-    #[test]
-    fn agents_updown_moves_focus_and_enter_opens_session() {
-        let model = model_with(&["a", "b", "c"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Agents;
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.focused, 2);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(ui.focused, 1);
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Session);
-    }
-
-    #[test]
-    fn focused_scope_gate_routes_decision_to_that_agent() {
-        let mut model = model_with(&["lead"]);
-        model.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::ScopeReview {
-                proposal: stella_protocol::ScopeProposal {
-                    summary: "big".into(),
-                    steps: vec![],
-                    estimated_files: 3,
-                    estimated_cost_usd: None,
-                },
-            },
-        });
-        let mut ui = ready_ui();
-        let action = handle_deck_key(ch('a'), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::ToAgent {
-                agent: "lead".into(),
-                input: UserInput::ScopeDecision(ScopeDecision::Approve),
-            })
-        );
-    }
-
-    #[test]
-    fn set_tab_stamps_the_switch_moment_only_on_change() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        assert!(ui.tab_switched_at.is_none(), "no motion before any switch");
-
-        // The key path routes through set_tab and stamps the moment.
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Agents);
-        let first = ui.tab_switched_at.expect("tab switch stamped");
-
-        // Switching to the SAME tab must not restart the sweep.
-        ui.set_tab(DeckTab::Agents);
-        assert_eq!(ui.tab_switched_at, Some(first));
-    }
-
-    #[test]
-    fn traces_filter_cycles_through_agents_and_back() {
-        let model = model_with(&["a", "b"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Traces;
-        assert_eq!(ui.trace_filter, None);
-        handle_deck_key(ch('f'), &model, &mut ui);
-        assert_eq!(ui.trace_filter.as_deref(), Some("a"));
-        handle_deck_key(ch('f'), &model, &mut ui);
-        assert_eq!(ui.trace_filter.as_deref(), Some("b"));
-        handle_deck_key(ch('f'), &model, &mut ui);
-        assert_eq!(ui.trace_filter, None);
-    }
-
-    fn model_with_queue(prompts: &[&str]) -> WorkspaceModel {
-        let mut m = model_with(&["lead"]);
-        for (i, p) in prompts.iter().enumerate() {
-            m.queue.enqueue((*p).to_string(), i as u64);
-        }
-        m
-    }
-
-    #[test]
-    fn bang_prefix_runs_a_shell_command_immediately_never_enqueued() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "!cargo build".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Shell("cargo build".into()));
-    }
-
-    #[test]
-    fn bang_prefix_only_strips_the_single_dispatch_marker() {
-        // The command text itself starts with `!` (e.g. `!important`), so
-        // the typed line is `!!important` — only the first `!` is the
-        // dispatch marker; the second belongs to the command.
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        for c in "!!important".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Shell("!important".into()));
-    }
-
-    #[test]
-    fn bang_prefix_beats_a_pending_ask_user_gate() {
-        let mut model = model_with(&["lead"]);
-        model.apply_inbound(&Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::AskUser {
-                id: "q1".into(),
-                question: "which db?".into(),
-                options: vec!["sqlite".into()],
-            },
-        });
-        let mut ui = ready_ui();
-        for c in "!ls".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Shell("ls".into()),
-            "a shell line is never swallowed as a gate answer"
-        );
-    }
-
-    #[test]
-    fn ctrl_t_toggles_the_queue_editor() {
-        let model = model_with_queue(&["one"]);
-        let mut ui = ready_ui();
-        assert!(!ui.queue_open);
-        handle_deck_key(ctrl('t'), &model, &mut ui);
-        assert!(ui.queue_open);
-        handle_deck_key(ctrl('t'), &model, &mut ui);
-        assert!(!ui.queue_open);
-    }
-
-    #[test]
-    fn up_arrow_on_session_opens_the_queue_editor_on_the_newest_prompt() {
-        let model = model_with_queue(&["first", "second", "third"]);
-        let mut ui = ready_ui();
-        assert_eq!(ui.tab, DeckTab::Session);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert!(ui.queue_open, "up from an empty composer opens the queue");
-        assert_eq!(ui.queue_sel, 2, "newest prompt selected");
-        // With no queue, Up falls back to transcript scrolling (no popup).
-        let empty = model_with(&["lead"]);
-        let mut ui2 = ready_ui();
-        handle_deck_key(key(KeyCode::Up), &empty, &mut ui2);
-        assert!(!ui2.queue_open);
-    }
-
-    #[test]
-    fn up_arrow_does_not_open_the_queue_editor_over_a_pasted_chip() {
-        // The live buffer is empty, but a pasted chip is still attached —
-        // editing a queued item would `Composer::load` and silently drop it,
-        // so the composer must not read as "empty" here.
-        let model = model_with_queue(&["first"]);
-        let mut ui = ready_ui();
-        // Force a chip regardless of the deck's (high) paste threshold — this
-        // test is about the chip interaction, not where the threshold sits.
-        ui.composer = crate::composer::Composer::with_paste_threshold(3);
-        ui.composer
-            .paste("line1\nline2\nline3\nline4\nline5\nline6");
-        assert!(ui.composer.buffer().is_empty());
-        assert!(
-            !ui.composer.is_empty(),
-            "the chip keeps the composer non-empty"
-        );
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert!(
-            !ui.queue_open,
-            "opening the queue editor here would drop the pasted chip on next edit"
-        );
-    }
-
-    #[test]
-    fn queue_editor_navigates_deletes_and_edits_as_a_list() {
-        let model = model_with_queue(&["first", "second", "third"]);
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('t'), &model, &mut ui);
-        // Navigate to the second prompt.
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.queue_sel, 1);
-        // ctrl+x deletes exactly the selected prompt.
-        assert_eq!(
-            handle_deck_key(ctrl('x'), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::QueueRemove { index: 1 })
-        );
-        // Enter pulls the selected prompt back into the composer for editing.
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::QueueRemove { index: 1 })
-        );
-        assert_eq!(ui.composer.buffer(), "second", "prompt loaded for editing");
-        assert!(!ui.queue_open, "editing returns to the composer");
-    }
-
-    #[test]
-    fn queue_clear_requires_two_ctrl_d_presses() {
-        let model = model_with_queue(&["a", "b"]);
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('t'), &model, &mut ui);
-        // First press only arms the confirm.
-        assert_eq!(
-            handle_deck_key(ctrl('d'), &model, &mut ui),
-            DeckAction::Handled
-        );
-        assert!(ui.queue_confirm_clear);
-        // Any other key disarms it.
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert!(!ui.queue_confirm_clear, "other keys disarm the confirm");
-        // Two consecutive presses clear.
-        handle_deck_key(ctrl('d'), &model, &mut ui);
-        assert_eq!(
-            handle_deck_key(ctrl('d'), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::QueueClear)
-        );
-        assert!(!ui.queue_open, "clearing closes the editor");
-    }
-
-    #[test]
-    fn ctrl_r_toggles_thinking_from_any_tab() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        assert!(!ui.thinking_expanded, "collapsed by default");
-        handle_deck_key(ctrl('r'), &model, &mut ui);
-        assert!(ui.thinking_expanded);
-    }
-
-    #[test]
-    fn deck_slash_popup_selects_completes_and_dispatches() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![
-            SlashCommand::new("/help", "show help"),
-            SlashCommand::new("/models", "list models"),
-        ];
-        handle_deck_key(ch('/'), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.slash_selected, 1);
-        // Tab completes into the buffer while the popup is open (it does NOT
-        // cycle tabs).
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        assert_eq!(ui.composer.buffer(), "/models");
-        assert_eq!(ui.tab, DeckTab::Session, "tab did not cycle");
-        // Enter dispatches the (still-matching) selection as a prompt.
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::Enqueue {
-                text: "/models".into()
-            })
-        );
-    }
-
-    #[test]
-    fn slash_files_switches_to_files_and_closes_an_open_diff() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![SlashCommand::new("/files", "files")];
-        ui.files_diff_open = true; // a diff was left open from a prior view
-        for c in "/files".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled, "/files is consumed locally");
-        assert_eq!(ui.tab, DeckTab::Files);
-        assert!(!ui.files_diff_open, "/files shows the tree, not a diff");
-    }
-
-    #[test]
-    fn slash_mcp_switches_to_the_mcp_tab() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![SlashCommand::new("/mcp", "mcp")];
-        for c in "/mcp".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled, "/mcp is consumed locally");
-        assert_eq!(ui.tab, DeckTab::Mcp);
-    }
-
-    #[test]
-    fn slash_mcp_search_jumps_straight_into_registry_search() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![
-            SlashCommand::new("/mcp", "mcp"),
-            SlashCommand::new("/mcp-search", "search the MCP registry"),
-        ];
-        for c in "/mcp-search".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled, "/mcp-search is deck-local");
-        assert_eq!(ui.tab, DeckTab::Mcp);
-        assert_eq!(ui.mcp.mode, crate::views::mcp::McpMode::Search);
-    }
-
-    #[test]
-    fn slash_on_the_mcp_tab_opens_the_command_menu_not_search() {
-        // The old `/`-enters-search trigger collided with the command menu;
-        // `/` must now behave on the MCP tab exactly as everywhere else —
-        // it starts a slash query in the composer.
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![SlashCommand::new("/mcp-search", "search")];
-        ui.set_tab(DeckTab::Mcp);
-        handle_deck_key(ch('/'), &model, &mut ui);
-        assert_eq!(
-            ui.mcp.mode,
-            crate::views::mcp::McpMode::Browse,
-            "`/` no longer enters MCP search"
-        );
-        assert_eq!(ui.composer.buffer(), "/", "the slash query is typing");
-        assert!(
-            !slash_matches(&ui).is_empty(),
-            "…and the command menu is open over it"
-        );
-    }
-
-    #[test]
-    fn mcp_tab_navigates_toggles_and_enters_search() {
-        use crate::envelope::McpServerInfo;
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.set_tab(DeckTab::Mcp);
-        ui.mcp.servers = vec![
-            McpServerInfo {
-                name: "github".into(),
-                kind: "http".into(),
-                enabled: true,
-                connected: true,
-                health: Some("live".into()),
-                tool_count: 3,
-                auth_fields: vec!["Authorization".into()],
-                oauth: Some(false),
-                calls: 5,
-            },
-            McpServerInfo {
-                name: "fs".into(),
-                kind: "stdio".into(),
-                enabled: true,
-                connected: false,
-                health: None,
-                tool_count: 0,
-                auth_fields: vec![],
-                oauth: None,
-                calls: 0,
-            },
-        ];
-        // ↓ moves the selection.
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.mcp.selected, 1);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(ui.mcp.selected, 0);
-
-        // `e` toggles the selected server (session enable/disable).
-        let action = handle_deck_key(ch('e'), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::McpToggle {
-                name: "github".into()
-            })
-        );
-
-        // `s` enters search mode; typing builds the query; Enter searches.
-        // (`/` no longer does — it belongs to the command menu everywhere.)
-        handle_deck_key(ch('s'), &model, &mut ui);
-        assert_eq!(ui.mcp.mode, crate::views::mcp::McpMode::Search);
-        for c in "git".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(ui.mcp.query, "git");
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::McpSearch {
-                query: "git".into()
-            })
-        );
-        assert!(ui.mcp.searching);
-        // Esc leaves search mode.
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert_eq!(ui.mcp.mode, crate::views::mcp::McpMode::Browse);
-    }
-
-    #[test]
-    fn mcp_auth_prompt_captures_a_masked_value_as_a_redacted_secret() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.set_tab(DeckTab::Mcp);
-        ui.mcp.servers = vec![crate::envelope::McpServerInfo {
-            name: "github".into(),
-            kind: "http".into(),
-            enabled: true,
-            connected: true,
-            health: Some("live".into()),
-            tool_count: 1,
-            auth_fields: vec![],
-            oauth: Some(false),
-            calls: 0,
-        }];
-        // `a` enters auth mode.
-        handle_deck_key(ch('a'), &model, &mut ui);
-        assert_eq!(ui.mcp.mode, crate::views::mcp::McpMode::Auth);
-        // Type the field name, Enter advances to the value step.
-        for c in "TOKEN".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.mcp.auth.step, crate::views::mcp::AuthStep::Value);
-        for c in "sk-secret".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        match action {
-            DeckAction::Send(WorkspaceInput::McpAuth {
-                server,
-                field,
-                value,
-            }) => {
-                assert_eq!(server, "github");
-                assert_eq!(field, "TOKEN");
-                assert_eq!(value.reveal(), "sk-secret");
-                // The secret never appears under Debug.
-                assert!(!format!("{value:?}").contains("sk-secret"));
-            }
-            other => panic!("expected McpAuth, got {other:?}"),
-        }
-        assert_eq!(ui.mcp.mode, crate::views::mcp::McpMode::Browse);
-    }
-
-    #[test]
-    fn a_pasted_secret_lands_in_the_credential_value_never_the_composer() {
-        // THE paste-routing security P1: a bracketed paste while the MCP auth
-        // VALUE step is focused used to fall through to the global composer, so
-        // a pasted API token landed — in plaintext — in the prompt buffer (and
-        // could be sent, or shown in the transcript). It must route to the
-        // credential value input instead.
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.set_tab(DeckTab::Mcp);
-        ui.mcp.servers = vec![crate::envelope::McpServerInfo {
-            name: "github".into(),
-            kind: "http".into(),
-            enabled: true,
-            connected: true,
-            health: Some("live".into()),
-            tool_count: 1,
-            auth_fields: vec![],
-            oauth: Some(false),
-            calls: 0,
-        }];
-        handle_deck_key(ch('a'), &model, &mut ui); // enter auth
-        for c in "TOKEN".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui); // → Value step
-        assert_eq!(ui.mcp.auth.step, crate::views::mcp::AuthStep::Value);
-
-        // Paste a multi-line secret blob.
-        ui.paste("sk-pasted-token\nextra");
-
-        // It went into the credential value (newlines dropped — a one-line
-        // field), and did NOT leak into the global composer.
-        assert_eq!(ui.mcp.auth.value, "sk-pasted-tokenextra");
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "a pasted secret must never reach the composer: {:?}",
-            ui.composer.buffer()
-        );
-    }
-
-    #[test]
-    fn a_paste_in_skills_search_types_into_the_query_not_the_composer() {
-        // The SKILLS tab is keyboard-owning: a paste in its search pane must
-        // build the query, exactly like typed characters do, never the composer.
-        // (paste() is a direct DeckUi method, so no WorkspaceModel is needed.)
-        let mut ui = ready_ui();
-        ui.set_tab(DeckTab::Skills);
-        ui.skills.focus = SkillsFocus::Search;
-
-        ui.paste("postgres\nmigrations");
-
-        assert_eq!(ui.skills.query, "postgresmigrations");
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "a skills-tab paste must not reach the composer: {:?}",
-            ui.composer.buffer()
-        );
-    }
-
-    #[test]
-    fn slash_diff_switches_to_files_and_opens_the_diff() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![SlashCommand::new("/diff", "diff")];
-        for c in "/diff".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Files);
-        assert!(ui.files_diff_open, "/diff opens the viewer");
-    }
-
-    #[test]
-    fn a_refreshed_graph_snapshot_updates_the_view_out_of_band() {
-        use crate::graph::{GraphNode, GraphSnapshot};
-        let mut model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        assert!(ui.graph.is_none());
-        let snapshot = GraphSnapshot {
-            focus: "src/lib.rs".into(),
-            nodes: vec![GraphNode {
-                label: "src/lib.rs".into(),
-                kind: "file".into(),
-                location: Some("src/lib.rs".into()),
-            }],
-            edges: vec![],
-            files: vec!["src/lib.rs".into()],
-        };
-        ingest_inbound(
-            &Inbound::GraphSnapshot(snapshot.clone()),
-            &mut model,
-            &mut ui,
-        );
-        assert_eq!(ui.graph.as_ref(), Some(&snapshot));
-    }
-
-    // ── Graph file picker ────────────────────────────────────────────────────
-
-    /// A three-file graph rooted on the busiest (`src/b.rs`), on the Graph tab.
-    fn ui_with_graph() -> DeckUi {
-        use crate::graph::{GraphNode, GraphSnapshot};
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Graph;
-        ui.graph = Some(GraphSnapshot {
-            focus: "src/b.rs".into(),
-            nodes: vec![GraphNode {
-                label: "src/b.rs".into(),
-                kind: "file".into(),
-                location: Some("src/b.rs".into()),
-            }],
-            edges: vec![],
-            files: vec!["src/a.rs".into(), "src/b.rs".into(), "src/c.rs".into()],
-        });
-        ui
-    }
-
-    // AGENTS tab: INSTALLED AGENTS pane
-
-    fn installed_entry(name: &str, version: u32) -> InstalledAgentEntry {
-        InstalledAgentEntry {
-            name: name.into(),
-            description: format!("about {name}"),
-            tools: Some(vec!["Read".into()]),
-            scope: AgentScope::Project,
-            source_path: format!("/ws/.stella/agents/{name}.md"),
-            version,
-            versions: (1..=version)
-                .map(|v| crate::envelope::AgentVersionInfo {
-                    version: v,
-                    label: String::new(),
-                })
-                .collect(),
-            content: format!("---\nname: {name}\n---\nbody of {name}"),
-        }
-    }
-
-    /// A ready deck on the AGENTS tab's INSTALLED pane with `entries` loaded.
-    fn installed_ui(entries: Vec<InstalledAgentEntry>) -> DeckUi {
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Agents;
-        ui.agents_pane = AgentsPane::Installed;
-        ui.installed.entries = entries;
-        ui.installed.loaded = true;
-        ui
-    }
-
-    #[test]
-    fn slash_opens_the_picker_defaulting_to_the_current_focus() {
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        let action = handle_deck_key(ch('/'), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled);
-        assert!(ui.graph_picker_open, "/ opens the picker on the Graph tab");
-        // Default selection is the busiest/focused file (index 1 = src/b.rs),
-        // the sensible default — not forced there, just pre-selected.
-        assert_eq!(ui.graph_picker_sel, 1);
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "/ did not leak into the prompt"
-        );
-    }
-
-    #[test]
-    fn agents_pane_arrows_switch_and_first_visit_asks_for_the_list() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Agents;
-        assert_eq!(ui.agents_pane, AgentsPane::Executions, "executions first");
-        // → switches to INSTALLED AGENTS; the unloaded list triggers one
-        // refresh request.
-        let action = handle_deck_key(key(KeyCode::Right), &model, &mut ui);
-        assert_eq!(ui.agents_pane, AgentsPane::Installed);
-        assert_eq!(action, DeckAction::Send(WorkspaceInput::AgentsRefresh));
-        // ← switches back; → again does NOT re-fetch (busy flag pending).
-        handle_deck_key(key(KeyCode::Left), &model, &mut ui);
-        assert_eq!(ui.agents_pane, AgentsPane::Executions);
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Right), &model, &mut ui),
-            DeckAction::Handled,
-            "no duplicate refresh while one is in flight"
-        );
-    }
-
-    #[test]
-    fn enter_also_opens_the_picker() {
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert!(ui.graph_picker_open);
-    }
-
-    #[test]
-    fn typing_filters_and_re_anchors_the_selection() {
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        open_graph_picker(&mut ui);
-        // Filter to just "a.rs" — one match, selection re-anchors to 0.
-        handle_deck_key(ch('a'), &model, &mut ui);
-        assert_eq!(ui.graph_picker_query, "a");
-        assert_eq!(ui.graph_picker_sel, 0);
-        let matches = ui
-            .graph
-            .as_ref()
-            .unwrap()
-            .matching_files(&ui.graph_picker_query);
-        assert_eq!(matches, vec!["src/a.rs"]);
-    }
-
-    #[test]
-    fn enter_in_the_picker_re_roots_on_the_selected_file() {
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        open_graph_picker(&mut ui);
-        // A multi-char needle (`c.rs`) narrows to exactly src/c.rs — a bare
-        // `c` would also match the shared `src/` prefix of every file.
-        for c in "c.rs".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::FocusGraphFile {
-                file: "src/c.rs".into()
-            }),
-            "Enter sends the re-root request for the filtered selection"
-        );
-        assert!(!ui.graph_picker_open, "the picker closes on selection");
-    }
-
-    #[test]
-    fn down_arrow_walks_the_filtered_matches_and_clamps() {
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        open_graph_picker(&mut ui);
-        ui.graph_picker_sel = 0;
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui); // past the end
-        assert_eq!(ui.graph_picker_sel, 2, "clamps to the last of three files");
-    }
-
-    #[test]
-    fn esc_closes_the_picker_without_re_rooting() {
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        open_graph_picker(&mut ui);
-        let action = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled);
-        assert!(!ui.graph_picker_open);
-    }
-
-    #[test]
-    fn the_picker_is_modal_over_the_composer() {
-        // A printable key while the picker is open filters — it must NOT type
-        // into the global composer (the queue-editor modality contract).
-        let model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        open_graph_picker(&mut ui);
-        handle_deck_key(ch('b'), &model, &mut ui);
-        assert_eq!(ui.graph_picker_query, "b");
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "filter keys never reach the composer"
-        );
-    }
-
-    #[test]
-    fn a_re_rooted_snapshot_resets_the_node_cursor() {
-        let mut model = model_with(&["lead"]);
-        let mut ui = ui_with_graph();
-        ui.graph_cursor = 5; // stale cursor from the previous neighborhood
-        use crate::graph::{GraphNode, GraphSnapshot};
-        let rerooted = GraphSnapshot {
-            focus: "src/a.rs".into(),
-            nodes: vec![GraphNode {
-                label: "src/a.rs".into(),
-                kind: "file".into(),
-                location: Some("src/a.rs".into()),
-            }],
-            edges: vec![],
-            files: vec!["src/a.rs".into(), "src/b.rs".into(), "src/c.rs".into()],
-        };
-        ingest_inbound(&Inbound::GraphSnapshot(rerooted), &mut model, &mut ui);
-        assert_eq!(ui.graph_cursor, 0, "the cursor lands on the new focus");
-    }
-
-    #[test]
-    fn the_picker_does_not_open_without_a_loaded_graph() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Graph; // no snapshot loaded
-        handle_deck_key(ch('/'), &model, &mut ui);
-        assert!(!ui.graph_picker_open, "nothing to pick from — stays closed");
-    }
-
-    #[test]
-    fn slash_agents_opens_the_tab_on_the_installed_pane() {
-        let model = model_with(&["lead"]);
-        let mut ui = ready_ui();
-        ui.slash_commands = vec![SlashCommand::new("/agents", "agents")];
-        for c in "/agents".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Agents, "/agents opens the Agents tab");
-        assert_eq!(ui.agents_pane, AgentsPane::Installed);
-        assert_eq!(action, DeckAction::Send(WorkspaceInput::AgentsRefresh));
-        assert!(ui.composer.buffer().is_empty(), "the composer cleared");
-    }
-
-    #[test]
-    fn agents_list_ingest_updates_the_panel_out_of_band_and_clamps() {
-        let mut model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![]);
-        ui.installed.sel = 5;
-        ui.installed.busy = true;
-        ingest_inbound(
-            &Inbound::AgentsList {
-                entries: vec![installed_entry("reviewer", 1)],
-                status: Some("saved".into()),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert_eq!(ui.installed.entries.len(), 1);
-        assert_eq!(ui.installed.sel, 0, "selection clamped to the new list");
-        assert!(!ui.installed.busy, "a fresh list completes the op");
-        assert_eq!(ui.installed.status.as_deref(), Some("saved"));
-        assert_eq!(
-            model.agents.len(),
-            1,
-            "the model fold ignores the out-of-band list"
-        );
-    }
-
-    #[test]
-    fn installed_enter_opens_the_editor_and_ctrl_s_saves_a_new_version() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![installed_entry("reviewer", 2)]);
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::Edit);
-        assert_eq!(
-            ui.installed.editor.buffer(),
-            "---\nname: reviewer\n---\nbody of reviewer",
-            "the editor holds the pinned version's content"
-        );
-        // Type at the end (the cursor loads at the end of the buffer), with
-        // a plain Enter inserting a newline — never submitting.
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        for c in "x".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(ctrl('s'), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::AgentSave {
-                name: "reviewer".into(),
-                scope: AgentScope::Project,
-                content: "---\nname: reviewer\n---\nbody of reviewer\nx".into(),
-            }),
-            "ctrl+s sends the edited content — the driver writes a NEW pinned version"
-        );
-        assert_eq!(ui.installed.mode, InstalledMode::Browse);
-        assert!(ui.installed.busy, "save shows the working state");
-    }
-
-    #[test]
-    fn editor_esc_discards_without_sending_and_typing_never_leaks() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![installed_entry("reviewer", 1)]);
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        for c in "abc".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "editor typing never reaches the global composer"
-        );
-        let action = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled, "no save is sent");
-        assert_eq!(ui.installed.mode, InstalledMode::Browse);
-        assert!(
-            ui.installed
-                .status
-                .as_deref()
-                .is_some_and(|s| s.contains("discarded")),
-            "{:?}",
-            ui.installed.status
-        );
-    }
-
-    #[test]
-    fn create_flow_describes_picks_scope_and_dispatches_the_llm_draft() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![]);
-        handle_deck_key(ch('n'), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::CreateDescribe);
-        for c in "reviews diffs".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(ui.installed.create_desc, "reviews diffs");
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            ui.installed.mode,
-            InstalledMode::CreateScope,
-            "⏎ advances to the scope picker (mirrors the skills install flow)"
-        );
-        // Default scope is project; ↓ flips to user, ↑ flips back.
-        assert_eq!(ui.installed.create_scope(), AgentScope::Project);
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.installed.create_scope(), AgentScope::User);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::AgentCreate {
-                description: "reviews diffs".into(),
-                scope: AgentScope::Project,
-            })
-        );
-        assert_eq!(ui.installed.mode, InstalledMode::Browse);
-        assert!(ui.installed.busy);
-    }
-
-    #[test]
-    fn create_flow_requires_a_description_and_esc_steps_back() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![]);
-        handle_deck_key(ch('n'), &model, &mut ui);
-        // Empty description: ⏎ refuses to advance.
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::CreateDescribe);
-        for c in "x".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::CreateScope);
-        // Esc from the scope picker returns to the description, not Browse.
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::CreateDescribe);
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::Browse);
-    }
-
-    #[test]
-    fn version_picker_pins_an_older_version_without_editing() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![installed_entry("reviewer", 3)]);
-        handle_deck_key(ch('v'), &model, &mut ui);
-        assert_eq!(ui.installed.mode, InstalledMode::PickVersion);
-        assert_eq!(
-            ui.installed.version_sel, 2,
-            "the picker opens on the pinned version"
-        );
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::AgentPin {
-                name: "reviewer".into(),
-                scope: AgentScope::Project,
-                version: 1,
-            }),
-            "⏎ re-pins — an AgentPin, never an AgentSave"
-        );
-        assert_eq!(ui.installed.mode, InstalledMode::Browse);
-    }
-
-    #[test]
-    fn version_picker_on_the_already_pinned_version_sends_nothing() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![installed_entry("reviewer", 2)]);
-        handle_deck_key(ch('v'), &model, &mut ui);
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(action, DeckAction::Handled, "re-pinning the pin is a no-op");
-        assert!(
-            ui.installed
-                .status
-                .as_deref()
-                .is_some_and(|s| s.contains("already")),
-            "{:?}",
-            ui.installed.status
-        );
-    }
-
-    #[test]
-    fn installed_browse_letter_verbs_type_when_the_composer_has_text() {
-        let model = model_with(&["lead"]);
-        let mut ui = installed_ui(vec![installed_entry("reviewer", 1)]);
-        handle_deck_key(ch('h'), &model, &mut ui);
-        handle_deck_key(ch('n'), &model, &mut ui);
-        assert_eq!(
-            ui.installed.mode,
-            InstalledMode::Browse,
-            "a typed `n` is prompt text, not the create verb"
-        );
-        assert_eq!(ui.composer.buffer(), "hn");
-    }
-
-    // ── SKILLS tab ──────────────────────────────────────────────────────────
-
-    // `SkillOp`, `SkillScope`, `SkillSearchHit`, `SkillsView` arrive via
-    // `use super::*`; only `SkillRow` is not imported at module scope.
-    use crate::envelope::SkillRow;
-
-    fn skills_ui() -> DeckUi {
-        let mut ui = ready_ui();
-        ui.tab = DeckTab::Skills;
-        ui
-    }
-
-    fn a_row(name: &str, scope: SkillScope, enabled: bool) -> SkillRow {
-        SkillRow {
-            scope,
-            name: name.to_string(),
-            description: "d".to_string(),
-            body: "b".to_string(),
-            origin: "workspace".to_string(),
-            enabled,
-            version: 1,
-            latest: 1,
-            removable: true,
-        }
-    }
-
-    #[test]
-    fn skills_search_pane_types_a_query_and_dispatches_a_search() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.focus = SkillsFocus::Search;
-        for c in "pdf tools".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert_eq!(ui.skills.query, "pdf tools", "typed into the query");
-        assert!(
-            ui.composer.is_empty(),
-            "the global composer never saw the keys"
-        );
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Search {
-                query: "pdf tools".into()
-            }))
-        );
-    }
-
-    #[test]
-    fn skills_enter_on_a_hit_opens_the_scope_prompt_then_installs_scoped() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.focus = SkillsFocus::Search;
-        ui.skills.hits = vec![SkillSearchHit {
-            id: "acme/auth".into(),
-            installs: "1.2K installs".into(),
-            installs_rank: 1200,
-            url: "https://skills.sh/acme/auth".into(),
-        }];
-        ui.skills.query = "auth".into();
-        ui.skills.query_dirty = false;
-
-        let a = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(a, DeckAction::Handled);
-        assert!(matches!(
-            ui.skills.prompt,
-            Some(SkillPrompt::Scope {
-                action: ScopeAction::Install { .. },
-                ..
-            })
-        ));
-        handle_deck_key(ch('u'), &model, &mut ui);
-        let a = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            a,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Install {
-                scope: SkillScope::User,
-                id: "acme/auth".into()
-            }))
-        );
-        assert!(ui.skills.prompt.is_none());
-    }
-
-    #[test]
-    fn skills_ctrl_o_on_installed_opens_preview_with_local_body() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        let mut r = a_row("sql-style", SkillScope::Project, true);
-        r.body = "# SQL Style\n\nUse lowercase keywords.".to_string();
-        ui.skills.view = SkillsView {
-            rows: vec![r],
-            status: None,
-            busy: false,
-        };
-        // ctrl+o must NOT toggle chain-of-thought on the SKILLS tab — it opens
-        // the preview, with the body on hand (no driver round-trip).
-        let a = handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert_eq!(a, DeckAction::Handled);
-        let preview = ui.skills.preview.as_ref().expect("preview opened");
-        assert_eq!(preview.pending, None, "installed body is local");
-        assert!(
-            preview.body.as_deref().unwrap().contains("SQL Style"),
-            "body carried from the row"
-        );
-    }
-
-    #[test]
-    fn skills_ctrl_o_on_search_hit_requests_preview_and_shows_loading() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.focus = SkillsFocus::Search;
-        ui.skills.hits = vec![SkillSearchHit {
-            id: "acme/auth@oauth".into(),
-            installs: "1.2K installs".into(),
-            installs_rank: 1200,
-            url: "https://skills.sh/acme/auth/oauth".into(),
-        }];
-        let a = handle_deck_key(ctrl('o'), &model, &mut ui);
-        assert_eq!(
-            a,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Preview {
-                id: "acme/auth@oauth".into()
-            }))
-        );
-        let preview = ui.skills.preview.as_ref().expect("preview opened");
-        assert_eq!(preview.pending.as_deref(), Some("acme/auth@oauth"));
-        assert_eq!(preview.body, None, "loading until the driver replies");
-    }
-
-    #[test]
-    fn skills_preview_ingest_fills_matching_id_ignores_stale_and_esc_closes() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.preview = Some(SkillPreview {
-            title: "acme/auth@oauth".into(),
-            subtitle: String::new(),
-            pending: Some("acme/auth@oauth".into()),
-            body: None,
-            scroll: 0,
-        });
-        // A reply for a DIFFERENT hit is dropped (stale / re-targeted).
-        ingest_inbound(
-            &Inbound::SkillPreview {
-                id: "other/skill@x".into(),
-                body: "wrong".into(),
-                status: None,
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert_eq!(ui.skills.preview.as_ref().unwrap().body, None, "stale drop");
-        // The matching reply fills the body and clears the pending marker.
-        ingest_inbound(
-            &Inbound::SkillPreview {
-                id: "acme/auth@oauth".into(),
-                body: "# OAuth\n\nbody".into(),
-                status: None,
-            },
-            &mut model,
-            &mut ui,
-        );
-        let preview = ui.skills.preview.as_ref().unwrap();
-        assert!(preview.body.as_deref().unwrap().contains("OAuth"));
-        assert_eq!(preview.pending, None);
-        // Esc closes the overlay.
-        let a = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert_eq!(a, DeckAction::Handled);
-        assert!(ui.skills.preview.is_none(), "esc closes the preview");
-    }
-
-    #[test]
-    fn skills_preview_scroll_keys_move_the_offset() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.preview = Some(SkillPreview {
-            title: "x".into(),
-            subtitle: String::new(),
-            pending: None,
-            body: Some("a\nb\nc".into()),
-            scroll: 0,
-        });
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.skills.preview.as_ref().unwrap().scroll, 1);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(ui.skills.preview.as_ref().unwrap().scroll, 0);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        assert_eq!(
-            ui.skills.preview.as_ref().unwrap().scroll,
-            0,
-            "clamped at 0"
-        );
-    }
-
-    #[test]
-    fn skills_space_toggles_and_two_ctrl_x_uninstalls() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.view = SkillsView {
-            rows: vec![a_row("sql-style", SkillScope::Project, true)],
-            status: None,
-            busy: false,
-        };
-        let a = handle_deck_key(ch(' '), &model, &mut ui);
-        assert_eq!(
-            a,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::SetEnabled {
-                scope: SkillScope::Project,
-                name: "sql-style".into(),
-                enabled: false
-            }))
-        );
-        assert!(!ui.skills.view.rows[0].enabled, "optimistic flip");
-
-        let a1 = handle_deck_key(ctrl('x'), &model, &mut ui);
-        assert_eq!(a1, DeckAction::Handled);
-        assert!(ui.skills.uninstall_armed);
-        let a2 = handle_deck_key(ctrl('x'), &model, &mut ui);
-        assert_eq!(
-            a2,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Uninstall {
-                scope: SkillScope::Project,
-                name: "sql-style".into()
-            }))
-        );
-    }
-
-    #[test]
-    fn skills_e_opens_edit_overlay_and_ctrl_s_saves() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        let mut r = a_row("sql-style", SkillScope::Project, true);
-        r.body = "old body".into();
-        ui.skills.view = SkillsView {
-            rows: vec![r],
-            status: None,
-            busy: false,
-        };
-        handle_deck_key(ch('e'), &model, &mut ui);
-        assert!(matches!(
-            ui.skills.prompt,
-            Some(SkillPrompt::Edit { ref buffer, .. }) if buffer == "old body"
-        ));
-        for c in " +more".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let a = handle_deck_key(ctrl('s'), &model, &mut ui);
-        assert_eq!(
-            a,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Edit {
-                scope: SkillScope::Project,
-                name: "sql-style".into(),
-                body: "old body +more".into(),
-            }))
-        );
-    }
-
-    #[test]
-    fn skills_manage_hotkeys_yield_to_a_nonempty_composer() {
-        // THE skills-key P1: the installed-pane manage hotkeys (space/e/p/n)
-        // were claimed unconditionally, so typing a prompt on the SKILLS tab was
-        // hijacked — 'n' opened the create flow, 'e' the edit overlay, space
-        // toggled a skill. They must honor the deck-wide "hotkeys only from an
-        // empty composer" contract and, mid-prompt, build the prompt instead.
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        let r = a_row("sql-style", SkillScope::Project, true);
-        ui.skills.view = SkillsView {
-            rows: vec![r],
-            status: None,
-            busy: false,
-        };
-
-        // 'r' is not a hotkey → it falls through to the composer, so the
-        // composer is now non-empty.
-        handle_deck_key(ch('r'), &model, &mut ui);
-        assert_eq!(ui.composer.buffer(), "r");
-
-        // Now the manage-hotkey characters type into the composer, not fire.
-        for c in "enp e".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert!(
-            ui.skills.prompt.is_none(),
-            "no manage overlay may open while a prompt is being typed"
-        );
-        assert!(
-            ui.skills.view.rows[0].enabled,
-            "space must not toggle the skill mid-prompt"
-        );
-        assert_eq!(ui.composer.buffer(), "renp e");
-
-        // From an EMPTY composer, 'e' still opens the edit overlay as designed —
-        // the gate only defers to a prompt in progress, it doesn't disable the
-        // hotkeys.
-        ui.composer.clear();
-        handle_deck_key(ch('e'), &model, &mut ui);
-        assert!(matches!(ui.skills.prompt, Some(SkillPrompt::Edit { .. })));
-    }
-
-    #[test]
-    fn skills_p_opens_pin_picker_and_enter_pins() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        let mut r = a_row("s", SkillScope::Project, true);
-        r.version = 3;
-        r.latest = 3;
-        ui.skills.view = SkillsView {
-            rows: vec![r],
-            status: None,
-            busy: false,
-        };
-        handle_deck_key(ch('p'), &model, &mut ui);
-        assert!(matches!(
-            ui.skills.prompt,
-            Some(SkillPrompt::Pin {
-                sel: 3,
-                latest: 3,
-                ..
-            })
-        ));
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-        let a = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            a,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Pin {
-                scope: SkillScope::Project,
-                name: "s".into(),
-                version: 1,
-            }))
-        );
-    }
-
-    #[test]
-    fn skills_n_creates_via_description_then_scope() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        handle_deck_key(ch('n'), &model, &mut ui);
-        assert!(matches!(
-            ui.skills.prompt,
-            Some(SkillPrompt::CreateDescription { .. })
-        ));
-        for c in "extract tables from pdfs".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let a = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(a, DeckAction::Handled);
-        assert!(matches!(
-            ui.skills.prompt,
-            Some(SkillPrompt::Scope {
-                action: ScopeAction::Create { .. },
-                ..
-            })
-        ));
-        let a = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            a,
-            DeckAction::Send(WorkspaceInput::Skill(SkillOp::Create {
-                scope: SkillScope::Project,
-                description: "extract tables from pdfs".into(),
-            }))
-        );
-    }
-
-    #[test]
-    fn skills_snapshot_ingest_updates_view_and_clears_searching() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        ui.skills.searching = true;
-        let view = SkillsView {
-            rows: vec![a_row("a", SkillScope::Project, true)],
-            status: Some("done".into()),
-            busy: false,
-        };
-        ingest_inbound(&Inbound::Skills(view), &mut model, &mut ui);
-        assert_eq!(ui.skills.view.rows.len(), 1);
-        assert!(!ui.skills.searching, "a fresh list clears the spinner");
-        assert_eq!(ui.skills.status.as_deref(), Some("done"));
-        assert!(
-            model.agents.is_empty(),
-            "model fold ignores skills snapshots"
-        );
-    }
-
-    #[test]
-    fn skills_tab_still_leaves_via_tab_key() {
-        let model = WorkspaceModel::new();
-        let mut ui = skills_ui();
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        // The MCP tab now sits after SKILLS in the cycle, so Tab leaves SKILLS
-        // for MCP (still proving SKILLS is not a dead end).
-        assert_eq!(ui.tab, DeckTab::Mcp, "Tab cycles Skills → Mcp");
-    }
-
-    // ── ISSUES tab ─────────────────────────────────────────────────────────
-
-    use crate::envelope::{EntityField, EntityHit, IssueAction, IssueRow};
-
-    fn issues_ui() -> DeckUi {
-        let mut ui = ready_ui();
-        ui.set_tab(DeckTab::Issues);
-        ui
-    }
-
-    fn a_issue(key: &str) -> IssueRow {
-        IssueRow {
-            key: key.to_string(),
-            title: format!("title of {key}"),
-            state: "open".into(),
-            labels: vec!["bug".into()],
-            assignee: Some("@octocat".into()),
-            url: format!("https://github.com/o/r/issues/{key}"),
-            updated_at: None,
-        }
-    }
-
-    fn a_hit(kind: &str, label: &str, insert: &str) -> EntityHit {
-        EntityHit {
-            kind: kind.into(),
-            label: label.into(),
-            description: format!("about {label}"),
-            insert: insert.into(),
-        }
-    }
-
-    /// Open the create form and move focus to `field`.
-    fn form_on(ui: &mut DeckUi, model: &WorkspaceModel, field: IssueField) {
-        handle_deck_key(ch('n'), model, ui);
-        assert_eq!(ui.issues.mode, IssuesMode::Create);
-        while ui.issues.form_field != field {
-            handle_deck_key(key(KeyCode::Tab), model, ui);
-        }
-    }
-
-    #[test]
-    fn issues_first_tab_visit_queues_a_refresh() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        ui.set_tab(DeckTab::Mcp); // ISSUES is Mcp's Tab successor
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        assert_eq!(ui.tab, DeckTab::Issues);
-        assert!(ui.issues.busy, "the first visit loads without a keypress");
-        assert!(matches!(
-            ui.pending_inputs.as_slice(),
-            [WorkspaceInput::IssuesRefresh {
-                query: None,
-                state: None,
-                seq: 1,
-            }]
-        ));
-        // A second visit does not re-fetch (busy / loaded gate).
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        handle_deck_key(key(KeyCode::BackTab), &model, &mut ui);
-        assert_eq!(ui.pending_inputs.len(), 1, "no duplicate refresh");
-    }
-
-    #[test]
-    fn issues_browse_keys_refresh_and_start_work() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        ui.issues.rows = vec![a_issue("#7")];
-        ui.issues.loaded = true;
-        let action = handle_deck_key(ch('r'), &model, &mut ui);
-        assert!(matches!(
-            action,
-            DeckAction::Send(WorkspaceInput::IssuesRefresh { query: None, .. })
-        ));
-        let action = handle_deck_key(ch('w'), &model, &mut ui);
-        match action {
-            DeckAction::Send(WorkspaceInput::IssueAct { key, action, .. }) => {
-                assert_eq!(key, "#7");
-                assert_eq!(action, IssueAction::StartWork);
-            }
-            other => panic!("expected IssueAct, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn issues_tracker_search_fires_on_enter_and_esc_returns() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        handle_deck_key(ch('/'), &model, &mut ui);
-        assert_eq!(ui.issues.mode, IssuesMode::SearchTracker);
-        for c in "flaky".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "search typing never reaches the composer"
-        );
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        match action {
-            DeckAction::Send(WorkspaceInput::IssuesRefresh { query, .. }) => {
-                assert_eq!(query.as_deref(), Some("flaky"));
-            }
-            other => panic!("expected IssuesRefresh, got {other:?}"),
-        }
-        assert_eq!(ui.issues.mode, IssuesMode::Browse);
-    }
-
-    #[test]
-    fn issues_comment_prompt_sends_the_act_for_the_selected_issue() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        ui.issues.rows = vec![a_issue("ENG-42")];
-        handle_deck_key(ch('c'), &model, &mut ui);
-        assert_eq!(ui.issues.mode, IssuesMode::Comment);
-        for c in "lgtm".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::IssueAct {
-                key: "ENG-42".into(),
-                action: IssueAction::Comment("lgtm".into()),
-                seq: 1,
-            })
-        );
-    }
-
-    #[test]
-    fn typeahead_opens_on_the_first_char_and_fires_per_keystroke() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        form_on(&mut ui, &model, IssueField::Assignee);
-        assert!(!ui.issues.typeahead.open(), "closed until the first char");
-
-        // `@` alone opens the popup and searches the EMPTY query (the
-        // backend lists all members for it).
-        handle_deck_key(ch('@'), &model, &mut ui);
-        assert!(ui.issues.typeahead.open(), "first char opens the popup");
-        assert!(matches!(
-            ui.pending_inputs.last(),
-            Some(WorkspaceInput::EntitySearch {
-                field: EntityField::Assignee,
-                query,
-                ..
-            }) if query.is_empty()
-        ));
-
-        // Every subsequent edit re-fires — insert and backspace alike.
-        handle_deck_key(ch('m'), &model, &mut ui);
-        assert!(matches!(
-            ui.pending_inputs.last(),
-            Some(WorkspaceInput::EntitySearch { query, .. }) if query == "m"
-        ));
-        handle_deck_key(key(KeyCode::Backspace), &model, &mut ui);
-        assert!(matches!(
-            ui.pending_inputs.last(),
-            Some(WorkspaceInput::EntitySearch { query, .. }) if query.is_empty()
-        ));
-        assert_eq!(ui.pending_inputs.len(), 3, "one request per keystroke");
-
-        // Deleting the last character closes the popup entirely.
-        handle_deck_key(key(KeyCode::Backspace), &model, &mut ui);
-        assert!(!ui.issues.typeahead.open(), "empty field ⇒ popup closed");
-    }
-
-    #[test]
-    fn typeahead_drops_stale_hits_and_applies_the_newest() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        form_on(&mut ui, &model, IssueField::Assignee);
-        handle_deck_key(ch('m'), &model, &mut ui); // seq 1
-        handle_deck_key(ch('a'), &model, &mut ui); // seq 2
-        let newest_seq = ui.issues.typeahead.seq;
-        let mut m = WorkspaceModel::new();
-
-        // The keystroke-1 reply lands late: stale, dropped.
-        ingest_inbound(
-            &Inbound::EntityHits {
-                field: EntityField::Assignee,
-                seq: newest_seq - 1,
-                query: "m".into(),
-                hits: vec![a_hit("Person", "stale", "@stale")],
-            },
-            &mut m,
-            &mut ui,
-        );
-        assert!(ui.issues.typeahead.hits.is_empty(), "stale reply dropped");
-
-        // The newest reply applies.
-        ingest_inbound(
-            &Inbound::EntityHits {
-                field: EntityField::Assignee,
-                seq: newest_seq,
-                query: "ma".into(),
-                hits: vec![a_hit("Person", "macanderson", "@macanderson")],
-            },
-            &mut m,
-            &mut ui,
-        );
-        assert_eq!(ui.issues.typeahead.hits.len(), 1);
-        assert!(!ui.issues.typeahead.loading);
-
-        // A reply for the WRONG field never lands either.
-        ingest_inbound(
-            &Inbound::EntityHits {
-                field: EntityField::Label,
-                seq: newest_seq + 5,
-                query: "ma".into(),
-                hits: vec![a_hit("Label", "major", "major")],
-            },
-            &mut m,
-            &mut ui,
-        );
-        assert_eq!(ui.issues.typeahead.hits[0].label, "macanderson");
-    }
-
-    #[test]
-    fn typeahead_enter_replaces_the_assignee_and_esc_keeps_typed_text() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        form_on(&mut ui, &model, IssueField::Assignee);
-        for c in "mac".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        ui.issues.typeahead.hits = vec![a_hit("Person", "macanderson", "@macanderson")];
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert_eq!(
-            ui.issues.form_assignee, "@macanderson",
-            "enter REPLACES the assignee field with the hit's insert"
-        );
-        assert!(!ui.issues.typeahead.open(), "picking closes the popup");
-        assert_eq!(ui.issues.mode, IssuesMode::Create, "still in the form");
-
-        // Esc with the popup open closes it but keeps the field text.
-        handle_deck_key(ch('x'), &model, &mut ui);
-        assert!(ui.issues.typeahead.open());
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert!(!ui.issues.typeahead.open());
-        assert_eq!(ui.issues.form_assignee, "@macandersonx", "text kept");
-        assert_eq!(ui.issues.mode, IssuesMode::Create, "form still open");
-    }
-
-    #[test]
-    fn typeahead_tab_appends_labels_comma_separated() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        form_on(&mut ui, &model, IssueField::Labels);
-        for c in "bug, urg".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        // The query is the segment being typed, not the whole field.
-        assert!(matches!(
-            ui.pending_inputs.last(),
-            Some(WorkspaceInput::EntitySearch {
-                field: EntityField::Label,
-                query,
-                ..
-            }) if query == "urg"
-        ));
-        ui.issues.typeahead.hits = vec![a_hit("Label", "urgent", "urgent")];
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        assert_eq!(
-            ui.issues.form_labels, "bug, urgent",
-            "the picked label replaces the partial segment, comma-appended"
-        );
-    }
-
-    #[test]
-    fn entity_query_and_insert_helpers_cover_both_fields() {
-        assert_eq!(entity_query(EntityField::Assignee, "@mac"), "mac");
-        assert_eq!(entity_query(EntityField::Assignee, "@"), "");
-        assert_eq!(entity_query(EntityField::Label, "bug, ur"), "ur");
-        assert_eq!(entity_query(EntityField::Label, "bug"), "bug");
-
-        let mut assignee = "mac".to_string();
-        apply_entity_insert(&mut assignee, EntityField::Assignee, "@macanderson");
-        assert_eq!(assignee, "@macanderson");
-
-        let mut labels = "ur".to_string();
-        apply_entity_insert(&mut labels, EntityField::Label, "urgent");
-        assert_eq!(labels, "urgent", "a lone partial segment is replaced");
-        labels.push_str(", b");
-        apply_entity_insert(&mut labels, EntityField::Label, "bug");
-        assert_eq!(labels, "urgent, bug");
-    }
-
-    #[test]
-    fn issue_form_ctrl_s_submits_the_parsed_fields() {
-        let model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        form_on(&mut ui, &model, IssueField::Title);
-        for c in "Fix the flake".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui); // → Body
-        assert_eq!(ui.issues.form_field, IssueField::Body);
-        for c in "line one".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui); // newline in body
-        for c in "line two".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui); // → Labels
-        for c in "bug".chars() {
-            handle_deck_key(ch(c), &model, &mut ui);
-        }
-        // Close the popup the typing opened, then Tab to Assignee.
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-        assert_eq!(ui.issues.form_field, IssueField::Assignee);
-
-        let action = handle_deck_key(ctrl('s'), &model, &mut ui);
-        assert_eq!(
-            action,
-            DeckAction::Send(WorkspaceInput::IssueCreate {
-                title: "Fix the flake".into(),
-                body: "line one\nline two".into(),
-                labels: vec!["bug".into()],
-                assignee: None,
-                seq: 4, // the three label keystrokes consumed seqs 1–3
-            }),
-        );
-        assert_eq!(ui.issues.mode, IssuesMode::Browse);
-        assert!(ui.issues.busy);
-    }
-
-    #[test]
-    fn issues_list_ingest_folds_rows_clears_busy_and_drops_stale() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        ui.issues.busy = true;
-        ui.issues.list_wait = 3;
-        ui.issues.sel = 9;
-
-        // A stale reply (an older request) is ignored outright.
-        ingest_inbound(
-            &Inbound::IssuesList {
-                seq: 2,
-                outcome: Ok(vec![a_issue("#1")]),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert!(ui.issues.rows.is_empty(), "stale list dropped");
-        assert!(ui.issues.busy, "…and the newer request is still awaited");
-
-        // The awaited reply folds in: rows, clamped selection, notice.
-        ingest_inbound(
-            &Inbound::IssuesList {
-                seq: 3,
-                outcome: Ok(vec![a_issue("#1"), a_issue("#2")]),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert_eq!(ui.issues.rows.len(), 2);
-        assert_eq!(ui.issues.sel, 1, "selection clamped to the new list");
-        assert!(!ui.issues.busy);
-        assert!(ui.issues.loaded);
-        assert_eq!(
-            model.agents.len(),
-            0,
-            "the model fold ignores the out-of-band list"
-        );
-
-        // An error outcome lands in the notice line (the no-tracker hint).
-        ui.issues.list_wait = 4;
-        ingest_inbound(
-            &Inbound::IssuesList {
-                seq: 4,
-                outcome: Err("no tracker connected — run `stella connect github`".into()),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert!(
-            ui.issues
-                .notice
-                .as_deref()
-                .is_some_and(|n| n.contains("no tracker connected")),
-            "{:?}",
-            ui.issues.notice
-        );
-    }
-
-    #[test]
-    fn issue_act_done_ingest_reports_the_outcome() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = issues_ui();
-        ui.issues.act_wait = 2;
-        ui.issues.busy = true;
-        ingest_inbound(
-            &Inbound::IssueActDone {
-                seq: 2,
-                key: "#7".into(),
-                outcome: Ok("created #7 — https://github.com/o/r/issues/7".into()),
-            },
-            &mut model,
-            &mut ui,
-        );
-        assert!(!ui.issues.busy);
-        assert!(
-            ui.issues
-                .notice
-                .as_deref()
-                .is_some_and(|n| n.contains("created #7")),
-            "{:?}",
-            ui.issues.notice
-        );
-    }
-
-    // ── Help overlay ───────────────────────────────────────────────────────
-
-    #[test]
-    fn question_mark_opens_help_overlay() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        assert!(!ui.help_open);
-        handle_deck_key(ch('?'), &model, &mut ui);
-        assert!(ui.help_open, "? opens the help overlay");
-    }
-
-    #[test]
-    fn show_help_inbound_opens_the_overlay_at_the_top() {
-        let mut model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        // Simulate a prior scroll so we can prove ShowHelp resets it.
-        ui.help_scroll.top = 42;
-        ui.help_scroll.follow = false;
-        ingest_inbound(&Inbound::ShowHelp, &mut model, &mut ui);
-        assert!(ui.help_open, "ShowHelp opens the overlay");
-        assert_eq!(ui.help_scroll.top, 0, "ShowHelp resets scroll to the top");
-        assert!(!ui.help_scroll.follow);
-    }
-
-    #[test]
-    fn help_overlay_scrolls_with_arrow_keys() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        ui.help_open = true;
-        ui.help_scroll.follow = false;
-        // Fake a tall document so scrolling is meaningful.
-        ui.metrics.help_total = 100;
-        ui.metrics.help_height = 10;
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        assert_eq!(ui.help_scroll.top, 1, "↓ scrolls down one line");
-        handle_deck_key(key(KeyCode::PageDown), &model, &mut ui);
-        assert!(ui.help_scroll.top > 1, "PageDown scrolls by a page");
-    }
-
-    #[test]
-    fn help_overlay_closes_with_esc_or_q() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        ui.help_open = true;
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert!(!ui.help_open, "Esc closes the help overlay");
-        // Re-open and close with q.
-        ui.help_open = true;
-        handle_deck_key(ch('q'), &model, &mut ui);
-        assert!(!ui.help_open, "q closes the help overlay");
-    }
-
-    #[test]
-    fn help_overlay_does_not_close_on_random_key() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        ui.help_open = true;
-        // A letter that isn't q or ? must NOT close the overlay (it used to —
-        // "any key closes" made the long content unreadable).
-        handle_deck_key(ch('x'), &model, &mut ui);
-        assert!(ui.help_open, "a random key does not close the overlay");
-    }
-
-    fn call(step: u64, call_seq: u64, role: &str) -> crate::envelope::RecordedCallInfo {
-        crate::envelope::RecordedCallInfo {
-            turn_instance: 0,
-            step,
-            call_seq,
-            call_role: role.into(),
-            provider: "anthropic".into(),
-            model: "opus".into(),
-            estimated_input_tokens: 100,
-        }
-    }
-
-    #[test]
-    fn ctrl_g_opens_inspect_and_asks_the_driver_for_the_call_index() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        assert!(!ui.inspect_open);
-        let action = handle_deck_key(ctrl('g'), &model, &mut ui);
-        assert!(ui.inspect_open, "⌃g opens the INSPECT overlay");
-        assert!(
-            matches!(action, DeckAction::Send(WorkspaceInput::InspectRefresh)),
-            "opening asks for a fresh index: {action:?}"
-        );
-    }
-
-    #[test]
-    fn enter_on_a_call_requests_that_calls_coordinate_not_the_first() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('g'), &model, &mut ui);
-        ingest_inbound(
-            &Inbound::RecordedCalls(vec![
-                call(0, 0, "worker"),
-                call(3, 1, "summarization"),
-                call(3, 0, "worker"),
-            ]),
-            &mut WorkspaceModel::new(),
-            &mut ui,
-        );
-        handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert!(
-            matches!(
-                action,
-                DeckAction::Send(WorkspaceInput::InspectCall {
-                    turn_instance: 0,
-                    step: 3,
-                    call_seq: 1,
-                })
-            ),
-            "⏎ reconstructs the SELECTED call — the summarizer here: {action:?}"
-        );
-        assert!(ui.inspect_pending, "the detail shows a pending state");
-    }
-
-    #[test]
-    fn enter_on_an_empty_index_is_a_no_op_not_a_request_for_call_zero() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('g'), &model, &mut ui);
-        let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-        assert!(
-            matches!(action, DeckAction::Handled),
-            "no calls recorded — ⏎ must not fabricate a coordinate: {action:?}"
-        );
-        assert!(!ui.inspect_pending);
-    }
-
-    #[test]
-    fn esc_steps_back_from_the_detail_to_the_list_before_closing() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('g'), &model, &mut ui);
-        ingest_inbound(
-            &Inbound::InspectedCall(Box::new(crate::envelope::InspectView {
-                call: call(0, 0, "worker"),
-                messages: vec![crate::envelope::InspectMessage {
-                    role: "system".into(),
-                    content: "You are Stella.".into(),
-                }],
-                verified: true,
-                unresolved: 0,
-                digest_mismatches: 0,
-            })),
-            &mut WorkspaceModel::new(),
-            &mut ui,
-        );
-        assert!(ui.inspect_view.is_some());
-        assert!(!ui.inspect_pending, "the answer clears the pending flag");
-
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert!(ui.inspect_view.is_none(), "esc returns to the call list");
-        assert!(ui.inspect_open, "…and does NOT close the overlay");
-
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-        assert!(!ui.inspect_open, "a second esc closes it");
-    }
-
-    #[test]
-    fn inspect_selection_cannot_run_past_the_last_call() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('g'), &model, &mut ui);
-        ingest_inbound(
-            &Inbound::RecordedCalls(vec![call(0, 0, "worker"), call(1, 0, "worker")]),
-            &mut WorkspaceModel::new(),
-            &mut ui,
-        );
-        for _ in 0..10 {
-            handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-        }
-        assert_eq!(
-            ui.inspect_sel, 1,
-            "the selection indexes a real row — render must be able to trust it"
-        );
-    }
-
-    #[test]
-    fn inspect_is_modal_and_does_not_leak_keys_to_the_composer() {
-        let model = WorkspaceModel::new();
-        let mut ui = ready_ui();
-        handle_deck_key(ctrl('g'), &model, &mut ui);
-        handle_deck_key(ch('x'), &model, &mut ui);
-        assert!(
-            ui.composer.buffer().is_empty(),
-            "a letter typed over the overlay must not reach the composer"
-        );
-        assert!(ui.inspect_open, "…and must not close it either");
-    }
-}
+mod tests;
